Refactored test structure slightly
